+++ R/AllGenerics.R
@@ -99,7 +99,8 @@ setGeneric("is.subset",
     function(x, y = NULL, proper = FALSE, sparse = FALSE, ...) standardGeneric("is.subset"))
 
 setGeneric("is.significant",
-    function(x, transactions, alpha = 0.01, adjust = "bonferroni") standardGeneric("is.significant"))
+    function(x, transactions, method = "fisher", 
+      alpha = 0.01, adjust = "bonferroni") standardGeneric("is.significant"))
 
 setGeneric("is.redundant",
     function(x, ...) standardGeneric("is.redundant"))