+++ NEWS.md
@@ -1,7 +1,8 @@
 # Changes in version 1.5-0.1 (xx/xx/2016)
 * FIX: Error message for sorting with an unknown interest measure.
 * FIX: abbreviate works now for rules correctly.
-
+* is.significant now supports in addition to Fisher's exact test,
+  the chi-squared test.
 
 # Changes in version 1.5-0 (09/23/2016)